Move tools to tools folder
