+++ packages/pwa-kit-mcp/src/server/server.js
@@ -10,7 +10,7 @@ import {StdioServerTransport} from '@modelcontextprotocol/sdk/server/stdio.js'
 
 import {z} from 'zod'
 import {CreateAppGuidelinesTool, CreateNewComponentTool, DeveloperGuidelinesTool} from '../utils'
-import {TestWithPlaywrightTool} from '../utils/run-site-test-tool'
+import {TestWithPlaywrightTool} from '../tools/run-site-test'
 
 // NOTE: This is a workaround to import JSON files as ES modules.
 // eslint-disable-next-line @typescript-eslint/no-var-requires
