+++ packages/pwa-kit-mcp/src/tools/pwa-create-app-guideline.js
@@ -6,7 +6,7 @@
  */
 
 // Project dependencies
-import {EmptyJsonSchema, getCreateAppCommand, isMonoRepo, runCommand} from './utils'
+import {EmptyJsonSchema, getCreateAppCommand, isMonoRepo, runCommand} from '../utils/utils'
 
 const CREATE_APP_COMMAND = getCreateAppCommand()
 const DISPLAY_PROGRAM_FLAG = '--displayProgram'
