+++ packages/pwa-kit-mcp/src/tools/pwa-developer-guideline.js
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: BSD-3-Clause
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-import {EmptyJsonSchema} from './utils'
+import {EmptyJsonSchema} from '../utils/utils'
 
 const guidelinesText = `# Salesforce Commerce Composable Storefront Development Guidelines
