+++ packages/pwa-kit-mcp/src/tools/pwa-developer-guideline.test.js
@@ -4,8 +4,8 @@
  * SPDX-License-Identifier: BSD-3-Clause
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-import DeveloperGuidelinesTool from './pwa-developer-guideline-tool'
-import {EmptyJsonSchema} from './utils'
+import DeveloperGuidelinesTool from './pwa-developer-guideline'
+import {EmptyJsonSchema} from '../utils/utils'
 
 describe('PWA Development Guidelines', () => {
     describe('DeveloperGuidelinesTool', () => {
